Исправлено: 1. Дублирование конфигурации. В базовом appsettings у тебя установлены такие же настройки, поэтому здесь секцию можно удалить. 2. На модели Person нет ни одного атрибута валидации модели, поэтому вызов ModelState.IsValid. 3. Рекомендую использовать стандартные типы для полей БД: long - идентификаторы, внешние ключи double/decimal - числа с плавающей точкой (decimal - если важна точность) int - целочисленные значение, которые точно туда поместятся, например, года у даты enum - для полей которые имеют четко определенное и неизменное количество значений (например, дни недели, либо состояния объекта) bool - логические значения DateTime/DatetimeOffset - для дат без времени/со временем (также для дат со временем рекомендуется хранить их в БД в UTC, чтобы можно было легко получить дату в таймзоне пользователя) string - текстовая информация Использование стандартных типов позволит избежать возможных проблем с мапингом данных при использовании EF.

+++ WebApplicationMustToHave/Models/CommonInterfaces.cs
@@ -8,7 +8,7 @@ namespace WebApplicationMustToHave.Models
         /// <summary>
         /// Уникальный идентификатор
         /// </summary>
-        public uint Id { get; set; }
+        public long Id { get; set; }
 
         /// <summary>
         /// Наименование
@@ -21,7 +21,7 @@ namespace WebApplicationMustToHave.Models
         /// <summary>
         /// Уникальный идентификатор
         /// </summary>
-        public uint Id { get; }
+        public long Id { get; }
 
         /// <summary>
         /// Получает строку-представление.
